refactor: move common logic outside of chain-specific context
